[ffx][trace] Replace a mutable out-arg
As pointed out previously, it's kind of gross. I think I've been
spending too much time writing C++ lately.

Instead of having the symbolize_event_record take a Vec<u8> and &mut
bool and returning a Vec<u8> which is either the same as what was passed
in or a modified one (in which case the out-arg would be set to true, it
takes a &[u8] and returns an Option<Vec<u8>> so that the caller can
determine if there's been any modification.

Change-Id: I40b83d089da995bb0af7ba1281e328d20e24ad8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233105
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>