[starnix][vfs] Create FsNodeRareData struct
The FsNode structure contains many fields that are very rarely used.
This CL creates an FsNodeRareData struct to hold those fields.

In a typical usage sceniario we have about 7 MiB of FsNodes, of which at
least 2.5 MiB are the WaitQueues inside if `flock_info` and
`record_locks`. This CL probably saves more memory than that, but it's
just easy to see those fields in the heap profile because they are
separate allocations.

This CL removes 85% of the WaitQueues in the entire Starnix kernel.

Change-Id: Icc537b8d5533cb44cf2e89acdd7ec5d9f781e0c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233008
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>