[assembly][bt] Add config for Hands Free
Add a configuration to enable the Hands Free profile
(bt-hfp-hands-free). This config shares the same codecs and controller
features as the Audio Gateway role.

Define an AIB for the component and include it in the build if specified
via assembly. This AIB is experimental and has not gone through security
review yet. Add it to userdebug and eng builds only.

NOTE: This CL does not enable it on any products.

Bug: 401064356
Test: fx build
Test: fx test src/lib/assembly

Change-Id: I16288b930588d0346a3880ab8f91d5ff1444d544
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1221429
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>
Reviewed-by: Marie Janssen <jamuraa@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>