[build] Move zircon sysroot/sdk to sysroot_sdk so its files aren't included by RBE for linking
The gen/zircon/public/sysroot folder is included as a whole directory
structure, no files should be in it that aren't part of the sysroot
itself, including the sdk atom that describes it.

The sometimes-present nature of the sdk atom files causes cache misses
as depending on when build actions run vs. the sdk atom creation ends
the files may or may not be present, changing the overall closure of
files that are part of the remote action.

This shifts the sdk atom from a subdir to a sibling dir, so that its
files are excluded from the operation.

Bug: b/405227511
Change-Id: I3c3d9c45fd6e2e9862bfbeeabe09020bffbfc716
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233610
Commit-Queue: Aaron Wood <aaronwood@google.com>
Reviewed-by: David Turner <digit@google.com>
API-Review: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: Hunter Freyer <hjfreyer@google.com>
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: David Dorwin <ddorwin@google.com>