[fx][build] Improved warnings when using Ninja target names.
To encourage developers to only use GN labels when invoking
`fx build` (and to prepare them for the use of Bazel labels
in the future), this CL improves `fx build` to print a warning
when a target argument is a Ninja output file name.

The warning includes the correct GN label / build arguments to
use. In the case where several GN labels all produce an output
file with the same name, the message is turned into an error,
listing all possible cases.

This new behavior is only triggered when the argument is
a file name (i.e. does not contain a directory separator,
or does not begin with // or @).

Ninja paths like "some/path" are still treated in the same
way, where an error is printed if it is not declared in the
Ninja build plan, or a warning with the right build arguments
is printed otherwise.

In other words:

```
$ fx build bazel_sdk_tests
WARNING: Use '//build/bazel/bazel_sdk:bazel_sdk_tests' instead of Ninja target 'bazel_sdk_tests'

$ fx build zbi
ERROR: Ambiguous Ninja target name 'zbi' matches several GN labels:
//sdk/fidl/zbi:zbi(//build/fidl:fidling)
//src/lib/assembly/zbi:zbi(//build/toolchain:host_x64)
//zircon/tools/zbi:zbi(//build/toolchain:host_x64)

$ fx build host_x64/zbi
WARNING: Use '--host //zircon/tools/zbi' instead of Ninja path 'host_x64/zbi'
```

- gn_ninja_outputs.py: Add new is_valid_target_name() and
  target_name_to_gn_labels() methods to the NinjaOutputsBase
  class and its derived classes.

- Change the behavior of the `build_args_to_gn_labels` command
  in build/api/client.py: The --allow-unknown is replaced with
  --allow-targets.

  Note that --allow-unknown is still supported by the
  `label_to_ninja_paths` command, which is also used by
  `fx build`.

- tools/devshell/build: Use --allow-targets instead of
  --allow-unknown when invoking the build_args_to_gn_labels
  command

Bug: 42084668
Change-Id: I459b5563ce28c971f9bb92526a3bca1a5e5976dc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233124
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>