[netstack3] Extract SeqRanges data struct
Extract the core of the Assembler struct into SeqRanges so we can reuse
it later for the SACK scoreboard. The core of this is to keep track of
disjoint ranges in an ordered fashion and this is symetrical for the
send side scoreboard.

Bug: 42078221
Change-Id: I8a11479ad559ca2d1bfdf3e33908a317a0a6eb0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233565
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>