[test_pilot] add run_test and helpers
This CL adds run_test, the function that actually runs the tests, and
some related helpful types. run_test functions in the sense that it
will actually run a test and its post-processors, but it does not yet
do the output processing work required of test-pilot. This will happen
in a future CL.

Helpers include:
- OutputDirectory, which owns the structure of the output directory and
generates paths on behalf of run_test.
- StdWriter, which is used to route stdout and stderr to files in the
output directory and to another writer (in this case, writers for
test-pilot's stdout and stderr streams, respectively). StdWriter also
deletes the file it creates if it is never written to.
- InvocationLog, which records the invocation of the test and post-
processors and which is serialized into a file in the output directory.

test-pilot's name has been changed from test_pilot.

Test: new unit tests

Bug: 327640651

Change-Id: I2ef5de615dbf3730eb9c99e660c0161f8bed9ada
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233585
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>