[symbolizer] Update build rules and paths
The primary change of this CL is to update the incorrect assumption
that in-tree, the SDK root is always the $root_build_dir. Since the
SDK may be built in a sub directory (as we are currently making
improvements to IDK construction and usage in-tree), the location
of `.symbol-index.json` should be calculated from the build directory
from the configuration context when in-tree, and fall back to the IDK structure otherwise.

This also adds a build rule to add `symbolizer` to
$root_build_dir/host-tools to be aligned with how other host tools are
located by `ffx` before the complete IDK is built.

Finally, there is an update to OWNERS to replace people no longer on the
team.

Change-Id: I75b2a5e8d4730716451e4f85b2750eea5f11337f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232765
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Jacob Rutherford <jruthe@google.com>