[ffx][sdk] Adding SdkRoot::HostTools
This sdk root is the "root of last resort", when no SDK manifest can
be found, and the sdk.root is not explicitly set.

This commonly happens in integration tests in-tree that are using `ffx`
and an isolate dir.

The behavior of the SDK when using and HostTools SDKRoot is that
get_host_tool() will look in the same directory as the current process,
which is most likely ffx).

Adding this SdkRoot type enables running integration tests without
having to build the entire IDK, which will make iterating on these
tests faster.

Change-Id: I2d6784371abe218aecc2bf14213da9ae71852769
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232964
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>