[net][ffx] Cleaning up ffx in conformance test
This cleans up the staging for using ffx in the network conformance
tests.

This is in preparation for streamlining the IDK dependencies in-tree
when using ffx in a test and remove the hard coded manifest used by
ffx in-tree.

This updated test depends on the "HostTool" SdkRoot in ffx. This means
there is no manifest used to configure the SDK, but relies on host tools
being located in the same directory as ffx. In-tree this  should be
`$root_build_dir/host-tools` e.g. out/default/host-tools.

Change-Id: I94331ea4bd5453c3d4c1e74b95810e4cca3ee45b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1231284
Reviewed-by: Tony Gong <gongt@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>