[packet-formats] Extract NDP option type consts
Sometimes we want to be able to refer to a known NDP option type as a u8
without needing to include it in the union of NdpOptions that are
necessarily parsed when parsing an NDP message using packet-formats.

To facilitate this, define an option_types mod with those consts, and
then use those consts in the definition of the protocol enum (which
requires some small changes to the `create_protocol_enum!` macro).

Bug: 397475289
Change-Id: I7d844665444dd96b47b2125814e97b48c8e2eac9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235464
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>