[storage] Use fuchsia_sync everywhere
Replace std::sync types with the equivalent fuchsia_sync types. Fuchsia
builds with panic=abort so we don't need to handle Mutex poisoning so we
always unwrap the result of locking a Mutex. fuchsia_sync doesn't do
Mutex poisoning and doesn't return a Result when locking a Mutex.
fuchsia_sync also provides MappedMutexGuards that are still experimental
in std::sync.

This is not expected to have an impact on performance or memory.

Change-Id: I11534c9940c8bbadb895f578ae00b26a79e91290
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1231904
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Chris Suter <csuter@google.com>