[kernel] Rename KTrace::GetInstance -> GetInternalState
The GetInstance function was actually returning a reference to the
internal::KTraceState, so rename the function as such. This will
reduce confusion when we create a static KTrace instance, which we
will need to support the new per-CPU buffer architecture.

Bug: 404539312
Change-Id: Ie37ef83bae5c31778dc7ddf8a45ae96bf3b02ef1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235926
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Corey Tabaka <eieio@google.com>