[fshost] Verify open success during migration recursive copy.
Without verifying this we can race in the tests because the file or
directory may not actually be created by the time we verify it. More
than that, src may be empty which means that we never interact with
the dst copy again and never verify that it was successfully created
before destroying the old filesystem.

Bug: 405230091
Change-Id: I9bb1324f04285bb65f449c76afe43cbc87f635b2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1234224
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>