[starnix] Skip invalid nexthop configuration packets
See the bug for more details, but the TL;DR is that Linux will silently
discard invalid nexthop configurations, whilst continuing to process
the valid configurations it may have received. This differs from the
rust-netlink library's handling of invalid configurations:

https://github.com/rust-netlink/netlink-packet-route/pull/153

I've modified our fork to behave closer to Linux, but the above change
was pulled down to stay as close as possible with the upstream source.

Fixed: 387579405
Test: Local validation with repro script in bug
Change-Id: Id158c9ab5a5c288c6cb9f3485c567bde6dd89144
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235150
Commit-Queue: Aaron Knobloch <aknobloch@google.com>
Reviewed-by: Jeff Martin <martinjeffrey@google.com>