[starnix] Containers have a structured config schema.
This change defines a set of configuration capabilities in
//src/starnix/containers/container_config.shard.cml which can be
offered to any Starnix container to configure its behavior. These
configuration capabilities can extend or override the container
definition in a container's package.

For an initial example, makes it possible to add extra feature strings
for a particular instantiation of a container. This will allow tests to
keep a shared container definition but to turn on particular features
of that container which are not needed by all tests that use the
container.

The syscall tests generate an additional component for each suite that
exposes the configuration values. In many cases it is possible to
define config values in the same component that offers the capabilities,
but here that would create a dependency cycle since the test component
also `use`s capabilities from the container.

Change-Id: Ief36e44ec75dcb705ecefdb2de49185b5a1daba2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235174
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>