[ffx][emu] Updating how to find mkfs-dosfs for uefi
This "installs" mkfs_docsfs into the host tools dir. This is
the standard place to locate host tools in-tree.

It also updates ffx_config::get_host_tool to return and error if
the file does not exist vs. returning the path.

This unblocks the uefi workflows, host tests are coming in a follow up CL.

Bug: 406159261
Change-Id: I966e7fb1caf318b3f7697a285f880e069a3edbc2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236070
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>