[ffxutil] Support strict bootloader boot
This updates `BootloaderBoot()` to support strict. It also
adds the ability for a specific invocation to provide extra
configs. If those are duplicates of "global" configs, the
strict command-builder will strip them out.

It also changes botanist to specify serial targets as
"serial:<serial>". This works even when not in strict mode.

Bug: 400797988
Change-Id: I8823487569e16c8a5478e520f94a8656053b32b9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233067
Reviewed-by: Ina Huh <ihuh@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>