[wlan][sme] Don't use MacSublayerSupport in SME
Refactor SME to no longer depend on MacSublayerSupport. Specifically,
this removes the functionality to intersect rates and capabilities for
the SoftMAC AP implementation, which is currently only used in tests.

This also removes the QueryMacSublayerSupport calls on startup in
both the FullMAC and SoftMAC MLMEs.

As a byproduct of this refactor, fuchsia.wlan.common/DataPlaneExtension is
no longer used by either MLME or SME and will be removed in a future
change.

This is part of the effort to cleanup driver_features in
fuchsia.wlan.common.

Bug: 305303708
Bug: 406220225
Test: Unit tests on CI/CQ
Test: Antlion quick tests on CI/CQ
Change-Id: I475e93702539c0753a439758632725af00dc30aa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235564
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: John Tan <johntan@google.com>