[assembly] Support AB slotted recoveries
If a product has A and B slots for recovery, the partitions config can
be updated to indicate this. When flashing the images, the recovery
zbi/vbmeta will be flashed to the A and B slots rather than the R slots.

To indicate this to assembly use:

  fuchsia_partition(
      name = "recovery_a",
      partition_name = "recovery_a",
      slot = SLOT.A,
      type = PARTITION_TYPE.RECOVERY_ZBI,
  )

Instead of:

  fuchsia_partition(
      name = "recovery_r",
      partition_name = "recovery_r",
      slot = SLOT.R,
      type = PARTITION_TYPE.ZBI,
  )

Change-Id: I755cdba0e7e1e207da359420f4a1a05ee1f4f15b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1233751
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>