[flatland] Fix buffer collection image import
Upon "reenabling" direct-to-display, I discovered that it wasn't
actually working, because the current code required the Flatland client
to maintain a reference to the imported buffer collection for as long as
the the image was being used to render. At a glance, this behavior seems
to date back to 2022, but this conflicts with more recent observations.

The fix is to validate the image against `display_imported_images_`.

Also, a couple cleanups to more tightly scope local variables.

Fixed: 377979329
Change-Id: I2d89648f124b198a8dcd059c4c850f6db4362e67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232464
Reviewed-by: Craig Stout <cstout@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>