[kernel] Introduce static KTrace instance
Modifies the KTrace singleton to expose a static KTrace instance.
This will allow the object to eventually store ktrace metadata
like categories, whether or not writes are enabled, process and
thread names, etc.

For now, this is a functional no-op. It just starts the process of
moving functionality out of the internal KTraceState in preparation
for the usage of per-CPU buffers.

Bug: 404539312
Change-Id: I06f904523eb80cf45dd52b7b4c19e2bdf38c58a7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236640
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>