[ffx][driver] Report nodes that are unbindable as such
When a node is owned by it's parent, reporting that it's unbound is
confusing. We can check whether the node ref is valid to determine
whether or not we should be able to bound by a driver.

Fixed: 396498971
Change-Id: I0cbbb1c9614471c48a700f074fd55cdabc9eb1cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236071
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Eaton <nickeaton@google.com>