[ffx][sdk] Update finding host tools not in the sdk
This changes the get_host_tool to only return paths of tools
in the SDK manifest. If there is no manifest, which happens in-tree
or in an isolated test environment, in this case the tool is looked for
in the same directory as the running binary (i.e. ffx).

This fixes a problem of blindly checking the sdk root for the host command
since there are binaries which are compiled for Fuchsia, not the host environment.

Bug: 406159261
Change-Id: Ie4a94e8286ee486ea548e832a182761ee6d06cac
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1237364
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>