[goldfish] Remove CreateBuffer2() support.
CreateBuffer2() render control command is used to create external
data buffers shared between Vulkan instances and non-Vulkan clients.

Currently this render control command is not supported on
gfxstream-based AEMU (FEMU). Given that this is not used by any
in-tree or out-of-tree clients, we decide to remove support for this
method for now to unblock the AEMU migration to release-gfxstream
builds.

Test: goldfish-unittests
Test: goldfish-control-device-test
Bug: 405344910
Bug: 406545904
Change-Id: Ie4cd6353dbee4f924fe884834ed5ab16984ba511
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1237584
Commit-Queue: Yilong Li <liyl@google.com>
Reviewed-by: Craig Stout <cstout@google.com>