[fxfs] Support PENDING_ACCESS_TIME_UPDATE
Implement support to persist atime. Fxfs is informed that the node has
been accessed when it receives a GetAttributes request with
PENDING_ACCESS_TIME_UPDATE. For that case, update atime to now if
atime <= ctime/mtime (or if it more than a day old).

Also, ctime (updated for status changes) would typically be updated for
GetAttributes but should not be updated if PENDING_ACCESS_TIME_UPDATE is
set and access time is the only thing being updated (atime update due to
a file access should not update ctime).

Bug: 401073280
Fixed: 401071741
Test: fx test fuchsia-pkg://fuchsia.com/fxfs-tests#meta/fxfs-platform-test.cm
Test: fx test storage_conformance_test
Change-Id: Ia7ce4e491772514dc942d24b329544701269d513
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235945
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>