[heapdump] Transfer the name and vaddr of ExecutableRegions too
These fields will be needed when we start using the
//src/developer/ffx/lib/symbolize library for symbolization
instead of delegating to pprof.

For now, the only effect of this CL is that the Mapping
entries in the generated protobuf will have the "filename"
populated. Note that pprof replaces it with the detected
path when it resolves a mapping through its build ID.

Change-Id: I085072cfa026635a19a59a070853b9ff7541ef70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1235585
Reviewed-by: Étienne J. Membrives <etiennej@google.com>
API-Review: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Fabio D'Urso <fdurso@google.com>