[build] Add new debug_symbols build API module.
Add a new build_symbols.json build API module used to list
debug symbol files and manifests for a given build configuration.

For now, the content of that files is missing:

- Entries for ELF binaries from prebuilt_package() targets.
- Entries for ELF binaries from Bazel-generated targets.

These issues will be addressed by future CLs on top of this
one. The goal is to modify artifactory to use this module
as the source of truth when deciding to upload debug symbols
to cloud storage. See associated bug for context.

This lists debug symbol files for all artifacts for the current
build configuration.

+ Add a `//build/api/client print_debug_symbols` command that
  can also resolve the build-id value for each entry, either
  using the elf_build_id_file entry, looking for .build-id/
  in the unstripped file path, or simply extracting the
  build-id value from the ELF notes section using
  //build/images/elfinfo.py.

Bug: 404208759

Change-Id: Id5ffd58c84a84c8fb7b15528ef260a0b0197575a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236524
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: David Turner <digit@google.com>
Reviewed-by: David Dorwin <ddorwin@google.com>