[utest] aarch32 max_load_address checking
This change fixes two oversights in max_load_address usage:

For MapSharedStorage() there is an off-by-one which would leave the
trailing byte of a mapping at the max_load_address inaccessible to the
restricted mode environment, 0x100000000 versus 0xffffffff.

For restricted blob allocation, the current search only ensures that the
starting address is under the limit. This change adds checking to ensure
the entire blob is able to fit under the max_load_address.

Bug: 389733983
Change-Id: I7f307c9a4438310d5ff5f36a818aaeb3eb7322ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1186537
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Will Drewry <drewry@google.com>