[overnet][usb] Switch from overnet-usb to usb-vsock
Use the usb-vsock driver, which implements `fuchsia.hardware.vsock`
instead of `fuchsia.hardware.overnet` to allow the host to interact with
physical devices over usb in the same way they would if they were in an
emulator, and allow for evolution of protocols thanks to vsock having
port numbers.

This also brings in the vsock service, which does increase the size on
eng and userdebug builds a bit.

This is a reland of fxr/1232186, which was reverted due to a conflict
produced when something else had to be reverted.

Change-Id: Ib9d607efe71fedb3696fe64b7c3bc087126ade26
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238138
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Cameron Finucane <eep@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Size-Review: Aaron Wood <aaronwood@google.com>