[starnix] Use zx::Counter for timers
The counter is meant to be decremented whenever an incoming message is
handled. The container can suspend anytime the counter is 0, and the
counter is incremented by the runner proxy each time it sends a message
into the container.

The counter is also set to 1 when created, to keep the system alive
until the first hanging get has been scheduled.

Bug: 399939990

Change-Id: Iebc6a20183ae389b3d20282ca88aec5df0634478
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238364
Reviewed-by: Vickie Cheng <vickiecheng@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>