[power][broker] Make the ElementID an integer instead of a string
We have enough information in the events list + topology (see element
lifecycle section @ go/fuchsia-power-broker-obs-improve) to be able to
always derive the event name. Therefore, we don't need to encode the
name in the element_id and can rely on a random number for it. This
change ensures that the alive element ids are unique, however we may
reuse element ids as elements are removed which means that the event
list may have element ids that refer to different elements and tools have to take that into account.

The core goal is to reduce space usage in our inspect instrumentation.

Fixed: 404631544
Change-Id: I23dbc219b9f6791b4ad26ed449962ab1c76a9838
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1232424
Reviewed-by: Kyle Gong <kgong@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>