[fs_management] fix component name when launching fs
When launching a dynamic filesystem in a collection, we add -relative to
the end of one of the two names because the difference is one has a
relative url and one hass an absolute url. Mistakenly they were swapped.

Change-Id: I106e998f3cea07df5d065628cb755ccb8fa84e21
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238425
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>