[fidl][python] Implement union, table, and zx.Handle
This change implements handles by compiling them to the int type the
same as the dynamic bindings. The complication with making this more
robust is that the fidl_codec wrapper expects all handle types to be an
integer instead of a more complex type. It would be a breaking change
for all tests that deal directly with handles to modify that assumption.

Now that all FIDL types except protocol are implemented, many
conformances tests are added so that fidlgen_python passes most of the
conformance test suite.

Bug: 355717613
Change-Id: I7f3846d5f450af269ec5685b8279b4352dea0f61
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1212724
Commit-Queue: Charles Celerier <chcl@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>