[kernel][vm] Remove freeing_owned_pages
Existing call sites sometimes have to 'lie' when freeing pages from a
hierarchy and claiming that the pages are not owned by this node. While
true, they are owned by some node, and should logically be freed by
that node. The argument put forward in these locations is that
hierarchies do not need to be correct, since they will never have a
page source handling free.

Instead of having call sites make this argument and pass a false value
for freeing_owned_pages, this change puts it forward as a global
property, that objects that have a page source handling free do not
have children, that can be checked and then generally assumed. Doing
this allows for removal of the flag passed to free and simplifies
call sites where it was otherwise ambiguous what 'ownership' meant.

Change-Id: I8c32c295293ae0d564ff1459adfa09a0c4d5d0c7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1236944
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>