[ffx] Allow VSock address types
This is a pretty bloody change. A lot of FFX code assumes that targets
are identifiable by addresses, and those addresses are IP addresses.

That said not much of substance should happen in this CL. It's mostly a
no-op until we start using vsock.

Change-Id: I019f48aa5f2b0673267369ccf83168771619637b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1210315
Reviewed-by: Steven Grady <slgrady@google.com>
API-Review: Ian McKellar <ianloic@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>